+++ src/main/java/lotto/domain/WinningLotto.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.domain;
 
 public class WinningLotto {
