+++ src/main/java/lotto/view/InputView.java
@@ -4,7 +4,6 @@ import java.io.PrintStream;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Scanner;
-import lotto.Lotto;
 
 public class InputView {
