+++ src/test/java/lotto/LottoNumberFactoryTest.java
@@ -6,6 +6,8 @@ import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
 import java.util.stream.Collectors;
+import lotto.domain.LottoNumber;
+import lotto.domain.LottoNumberFactory;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.RepeatedTest;
