+++ src/test/java/lotto/LottoWalletTest.java
@@ -5,6 +5,9 @@ import static org.assertj.core.api.Assertions.assertThat;
 import java.util.List;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
+import lotto.domain.Lotto;
+import lotto.domain.LottoNumber;
+import lotto.domain.LottoWallet;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
