+++ src/test/java/lotto/MoneyWalletTest.java
@@ -3,6 +3,7 @@ package lotto;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException;
 
+import lotto.domain.MoneyWallet;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
