+++ src/test/java/lotto/StatisticsReportTest.java
@@ -3,6 +3,11 @@ package lotto;
 import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
+import lotto.domain.Lotto;
+import lotto.domain.LottoNumber;
+import lotto.domain.LottoWallet;
+import lotto.domain.Prize;
+import lotto.domain.StatisticsReport;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
